chore: Update AwpClient constructor to handle user access rights
The AwpClient constructor in AwpClient.php has been updated to include a new parameter, $checkUserAccessRights. This parameter is used to determine whether to check user access rights or not. If $checkUserAccessRights is false or if the access token is available, the constructor will set the token. This change improves the handling of user access rights in the AwpClient class.